refactor(engine): extract grouped/core submodules
Split grouped/core.rs along three natural seams:
- component_predicates.rs: pure predicates over TemplateComponent
  and Reference, plus type-variant resolution
- sentence_initial.rs: bibliography + note-start capitalization
  context handling
- template_policy.rs: article-journal / anonymous-entry
  bibliography template filtering and rewriting

core.rs 1629 -> 1095 lines (-36% from original 1708).
No behavior change; pure structural move.
