+++ .beans/csl26-9f7g--rust-simplifyrefine-groupedcorers.md
@@ -5,7 +5,7 @@ status: in-progress
 type: task
 priority: normal
 created_at: 2026-05-17T00:42:49Z
-updated_at: 2026-05-17T00:48:19Z
+updated_at: 2026-05-17T07:44:00Z
 ---
 
 Coupled simplify + refine pass on grouped/core.rs (1708 lines, 5 clippy::too_many_arguments suppressions, 8+ overlong fns). Plan: ~/.claude/plans/look-for-opportunities-for-modular-bubble.md
@@ -13,7 +13,7 @@ Coupled simplify + refine pass on grouped/core.rs (1708 lines, 5 clippy::too_man
 ## Todo
 
 - [x] Commit 1: bundle params (GroupRenderParams + new TemplateRenderContext), remove 5 too_many_arguments allows
-- [ ] Commit 2: extract template_policy.rs, component_predicates.rs, sentence_initial.rs
+- [x] Commit 2: extract template_policy.rs, component_predicates.rs, sentence_initial.rs
 - [ ] Commit 3: split overlong functions in remaining core.rs
 - [ ] Verify quality-gate baseline + oracle no-diff on apa.csl
 - [ ] Open PR refactor/grouped-core-simplify-refine
