+++ crates/citum-engine/src/processor/rendering/grouped/core.rs
@@ -8,34 +8,14 @@ use super::super::{
     TemplateRenderRequest, find_grouping_component, get_variable_key, has_contributor_component,
     leading_group_affix, strip_author_component, strip_leading_group_affixes,
 };
+use super::component_predicates::{is_term_only_component, resolve_type_variant};
 use super::group_citation_items_by_author;
 use crate::error::ProcessorError;
 use crate::reference::Reference;
-use crate::render::bibliography::{append_rendered_component, component_starts_new_sentence};
-use crate::render::component::render_component_with_format;
 use crate::render::{ProcTemplate, ProcTemplateComponent};
 use crate::values::{ComponentValues, ProcHints, RenderContext, RenderOptions};
-use citum_schema::{
-    NoteStartTextCase,
-    locale::GeneralTerm,
-    options::ArticleJournalNoPageFallback,
-    options::titles::TextCase,
-    reference::NumOrStr,
-    template::{DateVariable, NumberVariable, SimpleVariable, TemplateComponent},
-};
+use citum_schema::template::TemplateComponent;
 use std::borrow::Cow;
-
-#[derive(Clone, Copy)]
-enum ArticleJournalBibliographyMode {
-    StandardDetail,
-    DoiFallback,
-}
-
-#[derive(Clone, Copy)]
-enum AnonymousEntryBibliographyMode {
-    ContainerLed,
-    SuppressPrintLike,
-}
 
 struct GroupRenderState<'a> {
     first_item: &'a crate::reference::CitationItem,
@@ -70,27 +50,6 @@ struct TemplateRenderContext<'a> {
     options: &'a RenderOptions<'a>,
     hint: &'a ProcHints,
     template_index: usize,
-}
-
-/// Returns the first type-variant template whose selector matches `ref_type`,
-/// or `None` if there are no variants or none match.
-fn resolve_type_variant<'a>(
-    type_variants: Option<
-        &'a indexmap::IndexMap<citum_schema::template::TypeSelector, citum_schema::TemplateVariant>,
-    >,
-    ref_type: &str,
-) -> Option<&'a [TemplateComponent]> {
-    let selector_candidates = aliased_type_selector_candidates(ref_type);
-    type_variants?.iter().find_map(|(selector, variant)| {
-        if selector_candidates
-            .iter()
-            .any(|candidate| selector.matches(candidate))
-        {
-            variant.as_template()
-        } else {
-            None
-        }
-    })
 }
 
 impl Renderer<'_> {
@@ -894,247 +853,6 @@ impl Renderer<'_> {
         }
     }
 
-    fn apply_sentence_initial_context<F>(
-        &self,
-        components: &mut [ProcTemplateComponent],
-        context: RenderContext,
-        note_start_text_case: Option<NoteStartTextCase>,
-    ) where
-        F: crate::render::format::OutputFormat<Output = String>,
-    {
-        match context {
-            RenderContext::Bibliography => {
-                self.apply_bibliography_sentence_initial_context::<F>(components);
-            }
-            RenderContext::Citation => {
-                self.apply_note_start_sentence_initial_context(components, note_start_text_case);
-            }
-        }
-    }
-
-    fn apply_bibliography_sentence_initial_context<F>(
-        &self,
-        components: &mut [ProcTemplateComponent],
-    ) where
-        F: crate::render::format::OutputFormat<Output = String>,
-    {
-        let punctuation_in_quote = components
-            .first()
-            .and_then(|component| component.config.as_ref())
-            .is_some_and(|config| config.punctuation_in_quote);
-        let default_separator = components
-            .first()
-            .and_then(|component| component.bibliography_config.as_ref())
-            .and_then(|bib| bib.separator.as_deref())
-            .unwrap_or(". ")
-            .to_string();
-
-        let mut entry_output = String::new();
-        for component in components.iter_mut() {
-            let rendered = render_component_with_format::<F>(component);
-            if rendered.is_empty() {
-                continue;
-            }
-
-            if component_starts_new_sentence(
-                &entry_output,
-                &rendered,
-                &default_separator,
-                punctuation_in_quote,
-            ) {
-                component.sentence_initial = true;
-                self.apply_sentence_initial_transform(component, None);
-            }
-
-            let rendered = render_component_with_format::<F>(component);
-            append_rendered_component(
-                &mut entry_output,
-                &rendered,
-                &default_separator,
-                punctuation_in_quote,
-            );
-        }
-    }
-
-    fn apply_note_start_sentence_initial_context(
-        &self,
-        components: &mut [ProcTemplateComponent],
-        note_start_text_case: Option<NoteStartTextCase>,
-    ) {
-        let Some(text_case) = note_start_text_case else {
-            return;
-        };
-
-        for component in components.iter_mut() {
-            if !self.is_note_start_term_component(component) {
-                continue;
-            }
-
-            component.sentence_initial = true;
-            self.apply_sentence_initial_transform(component, Some(text_case));
-            break;
-        }
-    }
-
-    fn apply_sentence_initial_transform(
-        &self,
-        component: &mut ProcTemplateComponent,
-        note_start_text_case: Option<NoteStartTextCase>,
-    ) {
-        if !component.sentence_initial {
-            return;
-        }
-
-        let locale = Some(self.locale.locale.as_str());
-        match &component.template_component {
-            TemplateComponent::Contributor(_) => {
-                let case =
-                    crate::values::text_case::resolve_text_case(TextCase::CapitalizeFirst, locale);
-                if let Some(prefix) = component.prefix.as_mut() {
-                    // Explicit template prefix (e.g. ". Translated by ") — capitalize it.
-                    *prefix = crate::values::text_case::apply_text_case(prefix, case);
-                } else {
-                    // No explicit prefix: the role label (e.g. "edited by ") is baked
-                    // into the rendered value.  Capitalize the first word so that
-                    // sentence-initial contributors read "Edited by …" not "edited by …".
-                    component.value =
-                        crate::values::text_case::apply_text_case(&component.value, case);
-                }
-            }
-            // Pre-formatted group components — capitalize the first word of the
-            // rendered group value (e.g. "edited by Smith" as first child).
-            TemplateComponent::Group(_) => {
-                let case =
-                    crate::values::text_case::resolve_text_case(TextCase::CapitalizeFirst, locale);
-                component.value = crate::values::text_case::apply_text_case(&component.value, case);
-            }
-            TemplateComponent::Term(_) if self.is_note_start_term_component(component) => {
-                if let Some(case) = note_start_text_case {
-                    component.value = crate::values::text_case::apply_note_start_text_case(
-                        &component.value,
-                        case,
-                        locale,
-                    );
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn is_note_start_term_component(&self, component: &ProcTemplateComponent) -> bool {
-        matches!(
-            &component.template_component,
-            TemplateComponent::Term(term) if term.term == GeneralTerm::Ibid
-        )
-    }
-
-    fn apply_article_journal_bibliography_policy<'a>(
-        &self,
-        reference: &Reference,
-        template: Cow<'a, [TemplateComponent]>,
-    ) -> Cow<'a, [TemplateComponent]> {
-        let Some(mode) = self.article_journal_bibliography_mode(reference) else {
-            return template;
-        };
-
-        if !article_journal_template_needs_filter(template.as_ref(), mode) {
-            return template;
-        }
-
-        Cow::Owned(filter_article_journal_template_components(
-            template.as_ref(),
-            mode,
-        ))
-    }
-
-    fn apply_anonymous_entry_bibliography_policy<'a>(
-        &self,
-        reference: &Reference,
-        template: Cow<'a, [TemplateComponent]>,
-    ) -> Option<Cow<'a, [TemplateComponent]>> {
-        let Some(mode) = self.anonymous_entry_bibliography_mode(reference, template.as_ref())
-        else {
-            return Some(template);
-        };
-
-        match mode {
-            AnonymousEntryBibliographyMode::ContainerLed => {
-                Some(Cow::Owned(rewrite_anonymous_entry_template(
-                    template.as_ref(),
-                    mode,
-                    reference.ref_type().as_str(),
-                    reference_has_doi(reference),
-                )))
-            }
-            AnonymousEntryBibliographyMode::SuppressPrintLike => None,
-        }
-    }
-
-    fn article_journal_bibliography_mode(
-        &self,
-        reference: &Reference,
-    ) -> Option<ArticleJournalBibliographyMode> {
-        if reference.ref_type() != "article-journal" {
-            return None;
-        }
-
-        let fallback = self
-            .bibliography_config
-            .as_ref()?
-            .article_journal
-            .as_ref()?
-            .no_page_fallback?;
-
-        match fallback {
-            ArticleJournalNoPageFallback::Doi => {
-                if reference_has_pages(reference) {
-                    Some(ArticleJournalBibliographyMode::StandardDetail)
-                } else if reference_has_doi(reference) {
-                    Some(ArticleJournalBibliographyMode::DoiFallback)
-                } else {
-                    None
-                }
-            }
-        }
-    }
-
-    fn anonymous_entry_bibliography_mode(
-        &self,
-        reference: &Reference,
-        template: &[TemplateComponent],
-    ) -> Option<AnonymousEntryBibliographyMode> {
-        if !matches!(
-            reference.ref_type().as_str(),
-            "entry-dictionary" | "entry-encyclopedia" | "chapter"
-        ) {
-            return None;
-        }
-
-        if reference.ref_type() == "chapter" && !template_has_dictionary_entry_shape(template) {
-            return None;
-        }
-
-        if self.reference_has_visible_author(reference) {
-            return None;
-        }
-
-        if !template_has_primary_title(template) || !template_has_parent_container_title(template) {
-            return None;
-        }
-
-        if reference_has_online_access(reference) {
-            Some(AnonymousEntryBibliographyMode::ContainerLed)
-        } else {
-            Some(AnonymousEntryBibliographyMode::SuppressPrintLike)
-        }
-    }
-
-    fn reference_has_visible_author(&self, reference: &Reference) -> bool {
-        reference
-            .author()
-            .is_some_and(|author| !self.resolve_contributor_names(&author).is_empty())
-    }
-
     fn build_template_render_hint(
         &self,
         reference: &Reference,
@@ -1383,258 +1101,6 @@ impl Renderer<'_> {
             item_request.note_start_text_case,
         );
         self.render_item_from_template_with_format::<F>(reference, request, item_request.delimiter)
-    }
-}
-
-fn filter_article_journal_template_components(
-    components: &[TemplateComponent],
-    mode: ArticleJournalBibliographyMode,
-) -> Vec<TemplateComponent> {
-    components
-        .iter()
-        .filter_map(|component| filter_article_journal_template_component(component, mode))
-        .collect()
-}
-
-fn filter_article_journal_template_component(
-    component: &TemplateComponent,
-    mode: ArticleJournalBibliographyMode,
-) -> Option<TemplateComponent> {
-    if should_suppress_article_journal_component(component, mode) {
-        return None;
-    }
-
-    match component {
-        TemplateComponent::Group(list) => {
-            let mut filtered = list.clone();
-            filtered.group = filter_article_journal_template_components(&list.group, mode);
-            (!filtered.group.is_empty()).then_some(TemplateComponent::Group(filtered))
-        }
-        _ => Some(component.clone()),
-    }
-}
-
-fn article_journal_template_needs_filter(
-    components: &[TemplateComponent],
-    mode: ArticleJournalBibliographyMode,
-) -> bool {
-    components
-        .iter()
-        .any(|component| article_journal_component_needs_filter(component, mode))
-}
-
-fn rewrite_anonymous_entry_template(
-    template: &[TemplateComponent],
-    mode: AnonymousEntryBibliographyMode,
-    ref_type: &str,
-    prefer_doi: bool,
-) -> Vec<TemplateComponent> {
-    match mode {
-        AnonymousEntryBibliographyMode::ContainerLed => {
-            let mut rewritten = Vec::new();
-
-            if let Some(container_title) =
-                find_preferred_parent_container_component(template, ref_type)
-            {
-                rewritten.push(container_title.clone());
-            }
-            if let Some(issued) = find_first_component(template, is_issued_date_component) {
-                rewritten.push(issued.clone());
-            }
-            if let Some(primary_title) = find_first_component(template, is_primary_title_component)
-            {
-                rewritten.push(primary_title.clone());
-            }
-            if let Some(volume) = find_first_component(template, is_volume_component) {
-                rewritten.push(volume.clone());
-            }
-
-            if prefer_doi {
-                if let Some(doi) = find_first_component(template, is_doi_component) {
-                    rewritten.push(doi.clone());
-                }
-            } else if let Some(url) = find_first_component(template, is_url_component) {
-                rewritten.push(url.clone());
-            }
-
-            if rewritten.is_empty() {
-                template.to_vec()
-            } else {
-                rewritten
-            }
-        }
-        AnonymousEntryBibliographyMode::SuppressPrintLike => template.to_vec(),
-    }
-}
-
-fn find_first_component(
-    template: &[TemplateComponent],
-    predicate: impl Fn(&TemplateComponent) -> bool,
-) -> Option<&TemplateComponent> {
-    template.iter().find(|component| predicate(component))
-}
-
-fn find_preferred_parent_container_component<'a>(
-    template: &'a [TemplateComponent],
-    ref_type: &str,
-) -> Option<&'a TemplateComponent> {
-    if matches!(
-        ref_type,
-        "chapter" | "entry-dictionary" | "entry-encyclopedia"
-    ) && let Some(parent_monograph) =
-        find_first_component(template, is_parent_monograph_title_component)
-    {
-        return Some(parent_monograph);
-    }
-
-    find_first_component(template, is_parent_container_title_component)
-}
-
-fn article_journal_component_needs_filter(
-    component: &TemplateComponent,
-    mode: ArticleJournalBibliographyMode,
-) -> bool {
-    if should_suppress_article_journal_component(component, mode) {
-        return true;
-    }
-
-    match component {
-        TemplateComponent::Group(group) => {
-            article_journal_template_needs_filter(&group.group, mode)
-        }
-        _ => false,
-    }
-}
-
-fn is_term_only_component(component: &TemplateComponent) -> bool {
-    match component {
-        TemplateComponent::Term(_) => true,
-        TemplateComponent::Group(group) => group.group.iter().all(is_term_only_component),
-        _ => false,
-    }
-}
-
-fn should_suppress_article_journal_component(
-    component: &TemplateComponent,
-    mode: ArticleJournalBibliographyMode,
-) -> bool {
-    match mode {
-        ArticleJournalBibliographyMode::StandardDetail => is_doi_component(component),
-        ArticleJournalBibliographyMode::DoiFallback => is_article_detail_component(component),
-    }
-}
-
-fn template_has_primary_title(template: &[TemplateComponent]) -> bool {
-    template.iter().any(is_primary_title_component)
-}
-
-fn template_has_parent_container_title(template: &[TemplateComponent]) -> bool {
-    template.iter().any(is_parent_container_title_component)
-}
-
-fn template_has_dictionary_entry_shape(template: &[TemplateComponent]) -> bool {
-    template.iter().any(|component| {
-        matches!(
-            component,
-            TemplateComponent::Variable(variable) if variable.variable == SimpleVariable::Version
-        )
-    })
-}
-
-fn is_primary_title_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Title(title)
-            if title.title == citum_schema::template::TitleType::Primary
-    )
-}
-
-fn is_parent_container_title_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Title(title)
-            if matches!(
-                title.title,
-                citum_schema::template::TitleType::ParentSerial
-                    | citum_schema::template::TitleType::ParentMonograph
-            )
-    )
-}
-
-fn is_parent_monograph_title_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Title(title)
-            if title.title == citum_schema::template::TitleType::ParentMonograph
-    )
-}
-
-fn is_issued_date_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Date(date) if date.date == DateVariable::Issued
-    )
-}
-
-fn is_volume_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Number(number) if number.number == NumberVariable::Volume
-    )
-}
-
-fn is_url_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Variable(variable) if variable.variable == SimpleVariable::Url
-    )
-}
-
-fn is_doi_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Variable(variable) if variable.variable == SimpleVariable::Doi
-    )
-}
-
-fn is_article_detail_component(component: &TemplateComponent) -> bool {
-    matches!(
-        component,
-        TemplateComponent::Date(date) if date.date == DateVariable::Issued
-    ) || matches!(
-        component,
-        TemplateComponent::Number(number)
-            if matches!(
-                number.number,
-                NumberVariable::Volume | NumberVariable::Issue | NumberVariable::Pages
-            )
-    )
-}
-
-fn reference_has_pages(reference: &Reference) -> bool {
-    match reference.pages() {
-        Some(NumOrStr::Str(pages)) => !pages.trim().is_empty(),
-        Some(NumOrStr::Number(_)) => true,
-        None => false,
-    }
-}
-
-fn reference_has_doi(reference: &Reference) -> bool {
-    reference.doi().is_some_and(|doi| !doi.trim().is_empty())
-}
-
-fn reference_has_url(reference: &Reference) -> bool {
-    reference.url().is_some()
-}
-
-fn reference_has_online_access(reference: &Reference) -> bool {
-    reference_has_doi(reference) || reference_has_url(reference)
-}
-
-fn aliased_type_selector_candidates(ref_type: &str) -> Vec<&str> {
-    match ref_type {
-        "chapter" => vec!["chapter", "entry-dictionary"],
-        _ => vec![ref_type],
     }
 }
